www-client/inox: move inox-patchset and icons to a dedicated repo
The changes is to follow Gentoo's Ebuild Maintenance guide:

"In general, the Gentoo repository should only be used for storing
.ebuild files, as well as any relatively small companion files,
such as patches or sample configuration files. These types of
files should be placed in the /files directory to keep the main
mycat/mypkg directory uncluttered. Exceptions to this rule are for
larger patch files (we recommend this for patches above 20KB) which
should be distributed as tarballs ... so that people do not waste
excessive amounts of bandwidth and hard drive space. Also, you
should not add binary (non-ASCII) files to the git tree."

Reference:
https://devmanual.gentoo.org/ebuild-maintenance/maintenance-tasks/index.html

Package-Manager: Portage-2.3.40, Repoman-2.3.9
Manifest-Sign-Key: 0x5010AD684AB2A4EE