dev-util/gitlab-runner: shouldn't perform checks during a binary merge
Reported-By: @ahrs

---
When emerging dev-util/gitlab-runner using a binary package
(for example from a binhost or just via --usepkgonly if you've
created a package using quickpkg) checks are still performed
to see if the portage user is a member of the docker group.

Since we're merging a pre-built binary package these checks
are unnecessary.
---

Closes #32