eclass: import golang-{base,vcs-snapshot}.eclass from main tree
and revbump to support EAPI 7

- Update copyright header
- Make use of BDEPEND if EAPI=7
- Requires at least Go v1.10.1

I don't like the idea of using custom eclasses, as surely it will
annoy users that don't want to install the overlay, but are
interested in importing some package(s) for their local overlay.

However, the bump to EAPI 7 is taking ages, so I will temporarily
maintain these eclasses until the ones in main tree supports EAPI 7.